LayoutTests/imported/w3c/web-platform-tests/fetch/metadata/form.https.sub-expected.txt





Harness Error (TIMEOUT), message = null

PASS localhost -> localhost:9443 iframe: forced
PASS localhost -> www.localhost:9443 iframe: forced
TIMEOUT localhost -> www.127.0.0.1:9443 iframe: forced Test timed out
PASS localhost -> localhost:9443 iframe: user-activated
PASS localhost -> www.localhost:9443 iframe: user-activated
TIMEOUT localhost -> www.127.0.0.1:9443 iframe: user-activated Test timed out
PASS localhost -> localhost:9443 iframe: user-activated: sec-fetch-dest
PASS localhost -> localhost:9443 iframe: user-activated: sec-fetch-mode
PASS localhost -> localhost:9443 iframe: user-activated: sec-fetch-site
FAIL localhost -> localhost:9443 iframe: user-activated: sec-fetch-user assert_equals: expected "?1" but got ""
PASS localhost -> localhost:9443 iframe: forced: sec-fetch-dest
PASS localhost -> localhost:9443 iframe: forced: sec-fetch-mode
PASS localhost -> localhost:9443 iframe: forced: sec-fetch-site
PASS localhost -> localhost:9443 iframe: forced: sec-fetch-user
PASS localhost -> www.localhost:9443 iframe: forced: sec-fetch-dest
PASS localhost -> www.localhost:9443 iframe: forced: sec-fetch-mode
FAIL localhost -> www.localhost:9443 iframe: forced: sec-fetch-site assert_equals: expected "same-site" but got "cross-site"
PASS localhost -> www.localhost:9443 iframe: forced: sec-fetch-user
PASS localhost -> www.localhost:9443 iframe: user-activated: sec-fetch-dest
PASS localhost -> www.localhost:9443 iframe: user-activated: sec-fetch-mode
FAIL localhost -> www.localhost:9443 iframe: user-activated: sec-fetch-site assert_equals: expected "same-site" but got "cross-site"
FAIL localhost -> www.localhost:9443 iframe: user-activated: sec-fetch-user assert_equals: expected "?1" but got ""
